+++ app/src/main/java/woowacourse/paint/view/PaintViewModel.kt
@@ -7,11 +7,11 @@ import androidx.lifecycle.ViewModel
 import woowacourse.paint.domain.BrushColor
 import woowacourse.paint.domain.BrushWidth
 import woowacourse.paint.domain.Lines
-import woowacourse.paint.view.model.Inks
+import woowacourse.paint.view.model.pen.ink.Inks
 import woowacourse.paint.view.model.mapper.LineMapper.toDomain
 import woowacourse.paint.view.model.mapper.LinesMapper.toModel
 import woowacourse.paint.view.model.pen.EllipsePen
-import woowacourse.paint.view.model.pen.Ink
+import woowacourse.paint.view.model.pen.ink.Ink
 import woowacourse.paint.view.model.pen.LinePen
 import woowacourse.paint.view.model.pen.Pen
 import woowacourse.paint.view.model.pen.RectPen
@@ -25,7 +25,7 @@ class PaintViewModel : ViewModel() {
         get() = Transformations.map(_lines) { it.toModel() }
 
     private val _pen: MutableLiveData<Pen> = MutableLiveData(
-        EllipsePen(onAddInk = ::addPen).apply {
+        EllipsePen(onAddInk = ::addInk).apply {
             setColor(color)
         }
     )
@@ -45,25 +45,25 @@ class PaintViewModel : ViewModel() {
     }
 
     fun changeToLinePen() {
-        _pen.value = LinePen(onAddInk = ::addPen).apply {
+        _pen.value = LinePen(onAddInk = ::addInk).apply {
             setColor(color)
             setStrokeWidth(strokeWidth)
         }
     }
 
     fun changeToRectPen() {
-        _pen.value = RectPen(onAddInk = ::addPen).apply {
+        _pen.value = RectPen(onAddInk = ::addInk).apply {
             setColor(color)
         }
     }
 
     fun changeToEllipsePen() {
-        _pen.value = EllipsePen(onAddInk = ::addPen).apply {
+        _pen.value = EllipsePen(onAddInk = ::addInk).apply {
             setColor(color)
         }
     }
 
-    private fun addPen(ink: Ink) {
+    private fun addInk(ink: Ink) {
         _lines.value = _lines.value?.add(ink.toDomain())
     }
 }
