+++ app/src/main/java/woowacourse/paint/view/model/pen/Pen.kt
@@ -1,5 +1,7 @@
 package woowacourse.paint.view.model.pen
 
+import woowacourse.paint.view.model.pen.ink.Ink
+
 interface Pen {
     val ink: Ink
     fun startPaint(pointX: Float, pointY: Float)
