+++ app/src/main/java/woowacourse/paint/view/model/pen/RectPen.kt
@@ -3,6 +3,7 @@ package woowacourse.paint.view.model.pen
 import android.graphics.Paint
 import android.graphics.Path
 import woowacourse.paint.graphics.AutoSortedRect
+import woowacourse.paint.view.model.pen.ink.Ink
 
 class RectPen(
     override val ink: Ink = Ink(
