+++ app/src/main/java/woowacourse/paint/view/model/pen/ink/Ink.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.view.model.pen
+package woowacourse.paint.view.model.pen.ink
 
 import android.graphics.Canvas
 import android.graphics.Paint
